% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_reports.R
\name{AnnualReports}
\alias{AnnualReports}
\title{Acquire listing of company annual reports.}
\usage{
AnnualReports(symbol, foreign = FALSE)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest.}

\item{foreign}{A logical vector indicating whether the company is domestic or foreign. \code{foreign = FALSE} by default.}
}
\description{
Extracts and displays listing of annual reports filed by a company in a data frame.
}
\examples{
AnnualReports("TSLA")
AnnualReports("BABA", foreign = TRUE)
}
