% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_info.R
\name{CompanyInfo}
\alias{CompanyInfo}
\title{Acquire basic company information.}
\usage{
CompanyInfo(symbol)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest.}
}
\description{
Extracts and displays basic information relating to a given company in a data frame.
}
\examples{
CompanyInfo("GOOG")
CompanyInfo("TSLA")
}
