\name{finreportr-package}
\alias{finreportr-package}
\alias{finreportr}
\docType{package}
\title{
\packageTitle{finreportr}
}
\description{
\packageDescription{finreportr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{finreportr}
\packageIndices{finreportr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{finreportr}

Maintainer: \packageMaintainer{finreportr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
