% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCashFlow.R
\name{GetCashFlow}
\alias{GetCashFlow}
\title{Acquire statement of cash flow.}
\usage{
GetCashFlow(symbol, year, annual, quarter)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company
of interest.}

\item{year}{A numeric vector specifying the year during which the annual
report was filed.}

\item{annual}{A logical vector indicating if annual reports should be pulled.
\code{annual = TRUE} by default.}

\item{quarter}{A logical vector indicating if quarterly reports should be
pulled. \code{quarter = TRUE} by default.}
}
\description{
Extracts and displays statement of cash flow from the annual report of a
given company. This functionality is only available for queries of cash flow
statements that belong to domestic company. Note that all data returned by
this function comes from the company's Form 10-K, not Form 10-K/A.
}
\examples{
\dontrun{
GetCashFlow("FB", 2016)
}
}
