% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{Reports}
\alias{Reports}
\title{Acquire listing of company annual reports.}
\usage{
Reports(symbol, annual = TRUE, quarter = TRUE)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company
of interest.}

\item{annual}{A logical vector indicating if annual reports should be pulled.
\code{annual = TRUE} by default.}

\item{quarter}{A logical vector indicating if quarterly reports should be
pulled. \code{quarter = TRUE} by default.}
}
\description{
Extracts and displays listing of annual reports filed by a
  company in a data frame.
}
\examples{
AnnualReports("TSLA")
AnnualReports("HTHIY", foreign = TRUE)
}
