% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statecodes}
\alias{statecodes}
\title{EDGAR State and Country Codes}
\format{
A data frame with 310 rows and 2 variables:
 \itemize{
  \item state: State or Country Code
  \item state.name: State or Country Name
}
}
\source{
\url{https://www.sec.gov/edgar/searchedgar/edgarstatecodes.htm}
}
\usage{
statecodes
}
\description{
A dataset containing state and country codes used in the SEC EDGAR database.
}
\keyword{datasets}
