% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finreportr.R
\docType{package}
\name{finreportr-package}
\alias{finreportr}
\alias{finreportr-package}
\title{finreportr: Financial Data from U.S. Securities and Exchange Commission}
\description{
Download and display company financial data from the U.S. Securities
    and Exchange Commission's EDGAR database. It contains a suite of functions with
    web scraping and XBRL parsing capabilities that allows users to extract data from EDGAR 
    in an automated and scalable manner. See <https://www.sec.gov/edgar/searchedgar/companysearch.html>
    for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sewardlee337/finreportr}
  \item Report bugs at \url{https://github.com/sewardlee337/finreportr/issues}
}

}
