% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance_sheet.R
\name{GetBalanceSheet}
\alias{GetBalanceSheet}
\title{Acquire balance sheet.}
\usage{
GetBalanceSheet(symbol, year)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest.}

\item{year}{A numeric vector specifying the year during which the annual report was filed.}
}
\description{
Extracts and displays balance sheet from the annual report of a given company. 
This functionality is only available for queries fo balance sheets that belong to domestic companies. 
Note that all data returned by this function comes from the company's Form 10-K, not Form 10-K/A.
}
\examples{
\dontrun{
GetBalanceSheet("FB", 2016)
}
}
