% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siccodes}
\alias{siccodes}
\title{Standard Industrial Classification Code List}
\format{
A data frame with 444 rows and 2 variables:
\itemize{
  \item SIC: Standard Industrial Classification Code
  \item industry: Industry Title
}
}
\source{
\url{https://www.sec.gov/info/edgar/siccodes.htm}
}
\usage{
siccodes
}
\description{
A dataset containing SIC codes and industries that they represent
}
\keyword{datasets}
